[1.3.6-prepare][Improvement-5275][Server] Move zk/ZKMasterClient and registry/ServerNodeManager into master
