+++ dolphinscheduler-server/src/main/java/org/apache/dolphinscheduler/server/master/zk/ZKMasterClient.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.dolphinscheduler.server.zk;
+package org.apache.dolphinscheduler.server.master.zk;
 
 import static org.apache.dolphinscheduler.common.Constants.SLEEP_TIME_MILLIS;
