+++ dolphinscheduler-server/src/main/java/org/apache/dolphinscheduler/server/worker/WorkerServer.java
@@ -53,9 +53,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
         @ComponentScan.Filter(type = FilterType.REGEX, pattern = {
                 "org.apache.dolphinscheduler.server.master.*",
                 "org.apache.dolphinscheduler.server.monitor.*",
-                "org.apache.dolphinscheduler.server.log.*",
-                "org.apache.dolphinscheduler.server.zk.ZKMasterClient",
-                "org.apache.dolphinscheduler.server.registry.ServerNodeManager"
+                "org.apache.dolphinscheduler.server.log.*"
         })
 })
 @EnableTransactionManagement
